requirements.txt
streamlit
google-generativeai
nltk